8282792: Loom: Fix optimized build after commit 4d2710
In commit 4d2710[1], one invocation of function print_frames() is moved
from condition `!PRODUCT` to condition `ASSERT`. As a result, all the
usages are now under condition `ASSERT`, leading to the compilation
warning "unused-function" in optimized VM build.

Hence, in this patch, we move the defintion of function print_frames()
to condition `ASSERT`.

[1] https://github.com/openjdk/loom/commit/4d2710#diff-74c5a41cb3476b29151b4a13732cf4756879de0aa1b1d8c4d73c38e570413fcfR2851